:recycle:refactor:자랑하기 컴포넌트 분리
